feat(assign): return type struct when assigning to type struct
adds support for extending type structs in a way that preserves the type behavior of the original
struct. Previous behavior would return an object struct with stricter requirements about unexpected
fields, rather than the loose requirements of the base type struct assigned "into".

BREAKING CHANGE: Behavior of struct returned by assign has changed

fix #720
